Add support for resampling 4 and 5 dimensional images
Created function decorator to extract 3d sub-volumes of images,
execute the function then reassemble the higher dimensional image.

The resample CLI make use of decorated functions to support moving
images of 5D. Also support for extracting the a 3D image from the
fixed image upon reading is added. This volume is only used for a
frame of reference.